feat: Add TransDice! config
TransDice is a spinoff server from TransPlace, not allowing people who
are not already members of TransPlace. Therefore, it makes sense to have
Theo in the server, but to strip out many of its bells and whistles

Several of the features of Theo are not needed by TransDice. An example
is that the welcoming feature is extraneous. Therefore, I have gated
some of the code behind checks to see if config is provided. If not, the
feature won't be used. I suspect some of this is written in ways that
are difficult enough to read that this might not be clear: it would be
worth doing some heavier refactoring around here, but I don't have the
time in this commit.

To keep track of this, I've written some typescript declarations for the
config. We don't currently use typescript in any capacity, but it might
be usable by a language server and it perhaps a bit more useful than the
documentation comment for precision...